Move dots interpolation to dots.c
